Added startup buttons and some styling to main screen.
activity_main.xml shows warning that either RelativeLayout or LinearLayout is useless as there are no child elements, but this can be ignored as some will likely be added later. If they are not added, and I stick with just the 3 buttons, then I can fix the error by removing RelativeLayout from the root element, making the LinearLayout loot, and replacing <android:layout_alignParentBottom="true"> with <android:layout_gravity="bottom">, and this will rid the warning.