Added camera activity startup when "Start Tracking" is pressed.
Added functionality such that when the "Start Recording" button is
pressed, a new activity called TrackingActivity.java is started, and this
opens a camera preview based on the color camera.

To-do: Add options to the camera preview such as changing to fish-eye
camera, changing ISO on color camera, adding record buttons, etc.
Finalizing user interface on mockup.