Created preliminary UI with popup menus.
Created a basic user interface consisting of record button, pause button,
flash settings, and manual/auto tracking selection. Still laying out the
buttons properly, and in the process of creating professional icons.

Also created a popup menu for the flash button such that when the flash
icon is clicked, user has option to turn flash On, Off or to Auto. The
menu is not yet functional.