Added select color option, refactored code.
Added a button that brings up a dialog to allow user to select a color for
overlays such as bounding boxes. I use a 3rd party library for this, which
can be found here: https://github.com/QuadFlask/colorpicker.

Also refactored onClick methods to make it easier to add multiple onClicks
for various button. I now implement OnClickActionListener, override
onClick(), and have a switch which calls external methods depending on
which view was clicked.