Wrote some code to record from surface preview.
Camera service dies regularly, so going to attempt to use MediaCodec. Last
resort issue is removing Tango support for now, and sticking to a regular
SurfaceView. The problem with camera service arises because of Tango
service completely taking over the camera when the service is connected,
so making calls to the camera API outside of Tango service will not work.