User touched positions now stored in YML file.
Decided to stop storing the video in Gallery. Instead, I opted to make a
folder in root called "ObjectTracker", and for each recording, make a
folder within "ObjectTracker", named with a timestamp (to avoid naming
clashes). This folder contains the recording itself, and the .yml file
with the touch information (time of touch, x and y coordinate of touch).