Fixed issues with output video being clipped.
Output video did not contain the full preview window when recorded, on the
galaxy s6. This was due to a scaling issue in CameraBridgeViewBase, so I
removed part of the code in deliverAndDrawFrame() that scales the bitmap
if mScale is > 0. Now the function draws a 1 to 1 corresponding bitmap to
the surface which is being recorded. This now results in a video that
records the full preview.

Also added javadocs to many methods and cleaned up some of the code.
Increased fps on the Nexus 5 by locking the auto exposure, and fixing the
range to 30.