Refactored code, increased fps during recording.
Refactored code to remove a lot of static variables. Changed videoSize for
MediaRecorder from getPreferredSize to 1280x720 (getPreferredSize returns
1920x720). Also changed in JavaCameraView's initializeCamera() function,
setPreviewSize to 1280x720. Reducing the size of the image means the
bitmap can be drawn quicker in deliverAndDrawFrame(). This is a quality
for higher fps tradeoff. Still looking into different ways to increase
fps.