Fixed some bugs with app crashing
App would crash when trying to record right after stopping a previous
recording. This occured because MediaRecorder was not being instantiated
again. so I removed the mediaRecorder instantation from constructor, and
put it in prepareMediaRecorder(). App would also crash when stopping
because of deliverAndDrawFrame() - mSurface would try to be accessed at
the same time it was released via releaseRecorder(), so added a lock to
both releaseRecorder() and deliverAndDrawFrame's mSurface access so that
only one can occur at a time. This should fix the crashing bugs.