Implemented some more template matching features.
Wrote some code that updates the template on each frame. The bounding box's contents are extracted and used as the template for the next frame. Also added some code that allows for manual correction of template; clicking on some location of the screen sets a new template, where the clicked location is the center point of the rectangle. The new template will be the same width as the original template.

Error checking also added. Test and cleanup/comment code now.