Added .yml file for automatic tracking also.
The .yml file is updated for each frame - append the coordinate of the center point of the rectangle. Because I use minMaxLoc which returns the top left coordinate, I have to add template width/2 and to the x component, and template height/2 to the y component, to get the center coordinate.

Also disabled the alpha slider when user selects automatic tracking. This is because the only way to add alpha to shapes is to use the "addWeighted" method which is computationally heavy, and I don't think it's worth having a transparent bounding box for a large fps hit.