[ISSUE #1] Allow to change modes during template initialization
Made an enhancement such that if template initialization procedure has started, the user now has the ability to cancel. This could be useful if automatic tracking was selected by accident, and the user wishes to change to manual. Previously, the user would have to set a template, record, and then cancel to be able to switch to manual.