Bug fixes with automatic tracking template correction
After adding the feature to extract regions from the last known location of object, template correction would not work properly because the last known location was not being updated in accordance to the new selected template position. This has now been fixed. Additionally, app was occasionally crashing because of a failed check in getROIFromImage() method (was checking width as opposed to height). This has also been fixed.