Changed manual tracking data file.
Sensor readings are now updated on every frame. Data is no longer appended in drawCircle() method - I simply update manualPosX and manualPosY, and write these values to the file on next frame. This means that the x and y positions of the object can be out by a frame. Not sure if this causes problems, but will think of ways to improve this.