Fixed drawRectangle() method. User can now choose to not record overlay.
Understanding of the drawRect() function was wrong - fixed and commented. Also added some code to cancel any previously showing toasts when showing a new toast. This was particularly noticeable when the user was selecting a template fast, as the toasts showed for a long lenth of time.

Made the onCameraFrame() method synchronize with the releaseMediaRecorder(). If these methods were not synchronized, onCameraFrame() could still draw a rectangle to the overlay even after releaseMediaRecorder() was called to clear the canvas, thus resulting in a stray rectangle that would not be cleared. This is because onCameraFrame() and releaseMediaRecorder() could ultimately run at the same time - now they can't.