Added debug activity to test sensors.
This new debug activity allows user to select which sensor they wish to
debug, and a number of iterations to poll that sensor for. The sensor data
is then appended to a yml file for the number of iterations, at the
fastest possible polling speed. This is useful for doing noise
evaluations, etc.